requirement.txt
discord.py
openai
python-dotenv
flask